add test for #994 + add option to return exceptional bool + refactor
To reproduce bug #944 in a test a device simulator needs to be able to
substitute bool value returned from hasPackageInstalled(). This has
been added to ExceptionSpec. In addition, some of the responsibilities
which previously belonged to the device simulator have been moved into
ExceptionSpec. Note that ExceptionSpec can either throw an exception or
return different bool, but not both.

Finally, some refactoring has been made.