make Exploration.run() return Failable instead of throwing exception
Any DeviceException thrown while Exploration.run() executed is returned,
 be it from the explorationLoop()/IApkExplorationOutput2 or from before
the loop.

This is the beginning of bigger rework to simplify collection of many
DeviceExceptions: now all the "try-" methods instead of throwing
DeviceExceptions, will return Failables with the exceptions or
aggregates of exceptions, to be cleanly handled in
ExploreCommand.execute().