handle apps crashing on launch during reset
(core change in RunnableResetAppExplorationAction)

Turns out, apps can crash in many ways on launch main activity:

1. the adb command hangs and the screen displays ANR. In such case, the
app will die as soon as the ANR is closed.
2. the adb command returns immediately and the screen displays ANR. The
 same behavior.

Now DroidMate determines if app crashed or not on launch, makes
appropriate assertions about app running or not and continues with
exploration. In any case, DroidMate ensures that home screen will be
displayed at the end of app reset action, or a DeviceException will
be thrown. Both cases are handled.