make TCP server return PID and package on connection check
Work in progress: so far implemented only on the device (TCP server)
side, not host (TCP client) side.

To return package name, now TCP server has reference to Context, which
is assigned when the monitor is initialized. I am not sure if it will be
always non-null, so null checks are inserted. Even if it is null, it
won't be a problem right, as currently it will be used only for logging
for manual bug diagnostics.