fix missing API logs
Previously only API logs containing package name of the logged app
in the stack trace were logged. Now I know that also if this is not
the case the log comes from the monitored app, because the logs come
only from processes which have initialized the inlined monitor, and
monitors were inlined only to the processes of the monitored app.

The fix means that also no log remapping is necessary, thus it got
deleted.