ensure the slf4j-simple dep. from utilities is properly excluded
Turns out "module" has to be specified in exclude to make the maven repo
pick it up, according to the official doc [1]:

"Gradle exclude rules are converted to Maven excludes if possible.
Such a conversion is possible if in the Gradle exclude rule the group
as well as the module name is specified (as Maven needs both in
contrast to Ivy). Per-configuration excludes are also included in
the Maven POM, if they are convertible."

[1] https://docs.gradle.org/2.12/userguide/maven_plugin.html#mappings