simplify DeviceModel implementation
now GuiState and UiautomatorWindowDump do not have a reference to
IDeviceModel. It was a conceptual error. Both GuiState and
UiautomatorWindowDump are "low-level raw" datatypes, while IDeviceModel
is an abstract concept, which cannot be referenced by such datatypes.
Furthermore, IDeviceModel is simple enough that just the necessary parts
of it can be passed in constructors to datatypes. For now it is just
androidLauncherPackageName and for testing, deviceDisplayDimensions.