remove deprecated code: legacy API lists.
Removed from MonitorGeneratorResources:
  final Path jellybeanPublishedApiMapping
  final Path jellybeanStaticMethods
  final Path appguardLegacyApis

These were the legacy API lists used for "Mining sandboxes" paper.
Following code, depending on these fields, was also removed:

ApiListStats, which was a quickly written class to provide API lists
stats. This class was constructed by a call to
"computeAndPrintApiListStats" in MonitorGeneratorFrontend. This
function was also removed.

ApiMapping, which held parsed versions of all th API lists. Note that
 this code contained methods to parse PScout API lists as well as
 "parseAppGuardLegacyApis" method. This last method had a comment
 explaining how the AppGuard APIs are parsed. The comment has been
 moved to the underlying ApiMethodSignature.fromDescription() method,
  as this call is being used to parse the current API list, also
  given in the "AppGuard APIs" format, which is actually just a jvm
  class format with the custom extension of "static" suffix.

The parameter "pn_appGuardOnlyApis" was removed from configuration.
It was used in ExplorationOutputDataExtractor.filterApiLogs to
include only app guard APIs. That made sense when the main API list
was PScout and had to be scoped to its AppGuard subset, but now we no
 longer have PScout API list. The main and only list is AppGuard now,
  which was since extended, as compared to the legacy AppGuard API
  list.