fix null socket bug
The problem was as follows: the MonitorServerRunnable runnable was
notifying the MonitorTcpServer.tryStart() before checking if exception
is null. Sometimes the server continued, calling next tryStart(),
assigning null both to the exception and to socket. Thus, the runnable
errorneously concluded that exception was always null and tried to
proceed with execution, encountering null socket.

Solution was to do the notify after the check if exeception was null,
not before.