fix: unnecessary attempts if device disconnected.
If device disconnected DroidMate will try to restart it. But before
doing it, it will try to reconnect ADB. Before this commit, this
would throw ADB exception (as the device is unavailable) that would be
propagated up to a function that would make DroidMate unnecessarily
retry operation again, without actually rebooting anything. Now the
aADB exception is ignored and reboot proceeds in spite of it.